pyproject.toml
[project]
name = "backstage_service_django"
version = "0.1.0"
description = ""
authors = [
    { name = "joe19921119", email = "bigjoe1992111@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "watchfiles (>=1.0.4,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "channels (>=4.2.0,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "coverage (>=7.6.10,<8.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-django (>=4.9.0,<5.0.0)",
    "pytest-factoryboy (>=2.7.0,<3.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "pillow (==11.1.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-spectacular-sidecar (>=2025.2.1,<2026.0.0)",
    "django-allauth (>=65.4.1,<66.0.0)",
    "django-cors-headers (==4.1.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "structlog (>=25.1.0,<26.0.0)",
    "django-structlog (>=9.0.1,<10.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "phonenumbers (>=8.13.55,<9.0.0)",
    "twilio (>=9.4.6,<10.0.0)",
    "sendgrid (>=6.11.0,<7.0.0)",
    "django-filter (>=25.1,<26.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
commitizen = "^4.2.1"
flake8 = "^7.1.2"
black = "^25.1.0"
isort = "^6.0.0"
pylint = "^3.3.4"
mypy = "^1.15.0"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true



[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311']


[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true


[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
